+++ R/grouped_ggpiestats.R
@@ -50,6 +50,7 @@
 grouped_ggpiestats <- function(data,
                                main,
                                condition = NULL,
+                               counts = NULL,
                                grouping.var,
                                title.prefix = "Group",
                                ratio = NULL,
@@ -69,6 +70,7 @@ grouped_ggpiestats <- function(data,
   # ================================ preparing dataframe ======================================
 
   if (!base::missing(condition)) {
+    if (base::missing(counts)) {
     # if condition variable *is* provided
     df <- dplyr::select(
       .data = data,
@@ -80,8 +82,22 @@ grouped_ggpiestats <- function(data,
         .data = .,
         title.text = !!rlang::enquo(grouping.var)
       )
+    } else {
+      df <- dplyr::select(
+        .data = data,
+        !!rlang::enquo(grouping.var),
+        !!rlang::enquo(main),
+        !!rlang::enquo(condition),
+        !!rlang::enquo(counts)
+      ) %>%
+        dplyr::mutate(
+          .data = .,
+          title.text = !!rlang::enquo(grouping.var)
+        )
+    }
   } else {
     # if condition variable is *not* provided
+    if (base::missing(counts)) {
     df <- dplyr::select(
       .data = data,
       !!rlang::enquo(grouping.var),
@@ -91,6 +107,18 @@ grouped_ggpiestats <- function(data,
         .data = .,
         title.text = !!rlang::enquo(grouping.var)
       )
+    } else {
+      df <- dplyr::select(
+        .data = data,
+        !!rlang::enquo(grouping.var),
+        !!rlang::enquo(main),
+        !!rlang::enquo(counts)
+      ) %>%
+        dplyr::mutate(
+          .data = .,
+          title.text = !!rlang::enquo(grouping.var)
+        )
+    }
   }
 
   # creating a nested dataframe
@@ -110,6 +138,7 @@ grouped_ggpiestats <- function(data,
     tidyr::nest(data = .)
 
   if (!base::missing(condition)) {
+    if (base::missing(counts)) {
     # creating a list of plots
     plotlist_purrr <- df %>%
       dplyr::mutate(
@@ -139,7 +168,39 @@ grouped_ggpiestats <- function(data,
             )
           )
       )
+    } else {
+      plotlist_purrr <- df %>%
+        dplyr::mutate(
+          .data = .,
+          plots = data %>%
+            purrr::set_names(!!rlang::enquo(grouping.var)) %>%
+            purrr::map(
+              .x = .,
+              .f = ~ggstatsplot::ggpiestats(
+                data = .,
+                main = !!rlang::enquo(main),
+                condition = !!rlang::enquo(condition),
+                counts = !!rlang::enquo(counts),
+                title = glue::glue("{title.prefix}: {as.character(.$title.text)}"),
+                ratio = ratio,
+                factor.levels = factor.levels,
+                stat.title = stat.title,
+                sample.size.label = sample.size.label,
+                caption = caption,
+                nboot = nboot,
+                palette = palette,
+                legend.title = legend.title,
+                facet.wrap.name = facet.wrap.name,
+                k = k,
+                facet.proptest = facet.proptest,
+                ggtheme = ggtheme,
+                messages = messages
+              )
+            )
+        )
+    }
   } else {
+    if (base::missing(counts)) {
     # creating a list of plots
     plotlist_purrr <- df %>%
       dplyr::mutate(
@@ -167,6 +228,34 @@ grouped_ggpiestats <- function(data,
             )
           )
       )
+    } else {
+      dplyr::mutate(
+        .data = .,
+        plots = data %>%
+          purrr::set_names(!!rlang::enquo(grouping.var)) %>%
+          purrr::map(
+            .x = .,
+            .f = ~ggstatsplot::ggpiestats(
+              data = .,
+              main = !!rlang::enquo(main),
+              counts = !!rlang::enquo(counts),
+              title = glue::glue("{title.prefix}: {as.character(.$title.text)}"),
+              ratio = ratio,
+              factor.levels = factor.levels,
+              stat.title = stat.title,
+              caption = caption,
+              nboot = nboot,
+              palette = palette,
+              legend.title = legend.title,
+              facet.wrap.name = facet.wrap.name,
+              k = k,
+              facet.proptest = facet.proptest,
+              ggtheme = ggtheme,
+              messages = messages
+            )
+          )
+      )
+    }
   }
 
   # combining the list of plots into a single plot